requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
alembic==1.17.2
python-dotenv==1.2.1
email-validator==2.3.0
requests==2.32.5
google-genai==1.61.0